% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_timepoints.R
\name{assign_timepoints}
\alias{assign_timepoints}
\title{Roughly categorise timepoints}
\usage{
assign_timepoints(obs_wide)
}
\description{
This may need to be finessed for each extraction.
}
\details{
Expects dataframe produced by \link{measures}.

Only does post intervention atm.
}
