% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_categories.R
\name{count_categories}
\alias{count_categories}
\title{Counts variables of interest}
\usage{
count_categories(netdat)
}
\arguments{
\item{netdat}{A netdat object (see visnetwork).}
}
\value{
Dataset of counts
}
\description{
How many observations do we have for each category?
}
\details{
How many studies?
}
