% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_categories.R
\name{count_gt}
\alias{count_gt}
\title{Basic summary count table}
\usage{
count_gt(count_gt_dat)
}
\arguments{
\item{count_gt_dat}{Data from \link{count_categories}.}
}
\value{
gt plot.
}
\description{
Generic output that will be the same for all studies at the first review.
}
