% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_obs.R
\name{extract_obs}
\alias{extract_obs}
\title{Extract observations}
\usage{
extract_obs(outcome, dat, cnames, measures_troubleshoot = FALSE)
}
\arguments{
\item{outcome}{Category from gsheet of outcome \link{outcome_keywords}.}

\item{dat}{\link{hpp_dat} data.}

\item{measures_troubleshoot}{For troubleshooting measures.}

\item{outcome_cols}{String of outcomes to select.}
}
\description{
Note that at the moment everything is included, but will need to update
the code for exclusions and combinations, as well.
}
\details{
Currently this is designed to follow from \link{preliminary_scrub} and output to
the \link{net_outcome} functions.

This does all the data handling that will happen on every outcome:
the objective is to perform the manipulations that are required for every
measure. Based on keywords, identify outcomes and massage into a long-format
meta-analysis dataset wherein each row signifies the reported statistics for
one arm of a study.

Best not to be too ambitious.
If the original dataset is also given as an argument to the next function,
any other covariates, etc. can be joined there. Only add to this function if
it is generalisable for all functions.
}
