% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_cols.R
\name{get_outcome_cols}
\alias{get_outcome_cols}
\title{Extract based on columns}
\usage{
get_outcome_cols(outcome, dat, outcome_cols)
}
\arguments{
\item{outcome}{pain or mood, etc.}

\item{dat}{hpp_dat in raw form}

\item{outcome_cols}{list of columns}
}
\description{
Extract based on columns
}
