% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpp_gt_pal.R
\name{hpp_gt_pal}
\alias{hpp_gt_pal}
\title{Table options for all tables}
\usage{
hpp_gt_pal(tab, first_col)
}
\arguments{
\item{tab}{A \link{gt} plot object.}

\item{first_col}{Name of the row-identifier column}
}
\description{
Sets colour of rowgroups, column header, background.
}
