% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_types.R
\name{measure_types}
\alias{measure_types}
\title{Check function for multiple measures}
\usage{
measure_types(dat)
}
\description{
This is for where there are multiple measures, otherwise,
I suspect, \link{measures_wide} won't run.
}
\details{
I also need to tag the measure, so, this can function as that as well
for now.
}
