% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_types.R
\name{measures_numeric}
\alias{measures_numeric}
\title{Convert measures to numeric}
\usage{
measures_numeric(outcome_obs)
}
\arguments{
\item{outcome_obs}{As created by [].}
}
\description{
This is failing so often it'd be useful to have a debugging algorithm
for it.
}
