% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_types.R
\name{measures_wide}
\alias{measures_wide}
\title{Identify measurements}
\usage{
measures_wide(obs_long, count_obs = FALSE)
}
\arguments{
\item{obs_long}{Long-form observations for an outcome.}

\item{count_obs}{Troubleshooting toggle that counts how many duplicates. Call
\link{extract_obs} with \code{measures_troubleshoot} to get the long-form expected.}
}
\description{
Ignoring counters at the end, assuming measure is at the end before possible
duplication tag, count the number and type (percent, n, limits, mean, sd,
etc.) from the end.
}
\details{
Assumes stringr::structure of the stringr::string:
\enumerate{
\item description
\item measure
\item possible duplication tag
}

Pivot wide by measure.
}
