% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_obs.R
\name{outcome_keywords}
\alias{outcome_keywords}
\title{Get keywords}
\usage{
outcome_keywords(outcome, keywords)
}
\arguments{
\item{outcome}{Specify which outcome from outcome column of keywords.}

\item{keywords}{Keywords dataset.}
}
\description{
Extract dataframe from \href{https://docs.google.com/spreadsheets/d/1vUY0gJ0bl-huVO-QIy7hjXwNRe87rPIt0cO4fLZobEQ/edit#gid=0}{googlesheet}
}
