% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise_wide}
\alias{pairwise_wide}
\title{Get wide-form pairwise comparisons}
\usage{
pairwise_wide(dat, trt)
}
\arguments{
\item{dat}{long-format outcome data}

\item{trt}{one treatment}
}
\description{
Get wide-form pairwise comparisons
}
