% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_hpp_dat.R
\name{raw_hpp_dat_path}
\alias{raw_hpp_dat_path}
\title{Get the latest raw data}
\usage{
raw_hpp_dat_path(now = lubridate::now())
}
\arguments{
\item{now}{this is just here so I can figure out how to update data in targets.}
}
\description{
Read from the data-raw file: latest version of the raw data exported from
Covidence.
}
